+++ __tests__/setup.js
@@ -1,0 +1,3 @@
+const chalk = require('chalk');
+
+chalk.level = 2;